rename layout folder to views and organize react pages
