--- client/bash.exe.stackdump
@@ -1,9 +1,0 @@
-Stack trace:
-Frame        Function    Args
-000FFFFA318  0018006133E (0018026DF0D, 0018024E186, 000FFFFA318, 000FFFF9210)
-000FFFFA318  00180049229 (00000000000, 00000000000, 00000000000, 00000000000)
-000FFFFA318  00180049262 (0018026DFC9, 000FFFFA1C8, 000FFFFA318, 00000000000)
-000FFFFA318  001800B5C18 (00000000000, 00000000000, 00000000000, 00000000000)
-000FFFFA318  001800B5D9D (000FFFFA330, 00000000000, 00000000000, 00000000000)
-000FFFFA5E0  001800B73CC (000FFFFA330, 00000000000, 00000000000, 00000000000)
-End of stack trace
