--- client/src/components/CreateProfile.js
@@ -1,3 +1,0 @@
-import { ReactComponent } from "*.svg";
-
-ReactComponent
