--- client/src/components/layout/Discussion.js
@@ -1,13 +1,0 @@
-import React, { Component } from "react";
-import Post from './Post';
-import PostEditor  from './PostEditor';
-
-export class DiscussionBoard extends Component {
-
-	render(){
-		return {
-
-		};
-	}
-}
- export default DiscussionBoard;
