+++ client/src/components/views/Auth/Login.js
@@ -2,7 +2,7 @@ import React, { Fragment, useState } from "react";
 import axios from "axios";
 import { connect } from "react-redux";
 import PropTypes from "prop-types";
-import { login } from "../../actions/auth";
+import { login } from "../../../actions/auth";
 import { Link, Redirect } from "react-router-dom";
 
 const Login = ({ login, isAuthenticated }) => {
