+++ client/src/components/views/Course/Course.js
