+++ client/src/components/views/Instructor/CreateInstructorProfile.js
@@ -1,7 +1,7 @@
 import React, { Fragment, useState } from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import { createProfile } from "../../actions/profile";
+import { createProfile } from "../../../actions/profile";
 import { Link, withRouter } from "react-router-dom";
 
 const CreateProfile = ({ createProfile, history }) => {
