+++ client/src/components/views/Lecture/ChatCard.js
