+++ client/src/components/views/LoggedOut.js
