+++ client/src/components/views/Student/StudentCourses.js
@@ -1,26 +1,32 @@
 import React, { Fragment } from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import { InstructorCourseCard } from "../InstructorCourseCard";
+import { CourseCard } from "./CourseCard";
+import { Button } from "react-bootstrap";
+import { Link } from "react-router-dom";
 
-const InstructorCourses = ({ course }) => {
+const StudentCourses = ({ course }) => {
   var max = 3;
   var count = 0;
   function func(number) {
     console.log(number);
-    return <InstructorCourseCard inputValue={course} />;
+    return <CourseCard inputValue={course} />;
   }
   const courses = course.map(
     course => (
       (count += 1),
       console.log(count),
       func(count),
-      (<InstructorCourseCard inputValue={course} />)
+      (<CourseCard inputValue={course} />)
     )
   ); //create a CourseCard for every course id in the Course array in Student P
 
   return (
     <Fragment>
+      <h3>Courses</h3>{" "}
+      <Link to="/enroll">
+        <Button variant="outline-primary">Enroll Course</Button>
+      </Link>
       <div className="courses">
         <div className="row">
           {courses}
@@ -32,8 +38,8 @@ const InstructorCourses = ({ course }) => {
   );
 };
 
-InstructorCourses.propTypes = {
+StudentCourses.propTypes = {
   course: PropTypes.array.isRequired
 };
 
-export default InstructorCourses;
+export default StudentCourses;
