+++ client/src/components/views/Student/StudentProfile.js
@@ -1,7 +1,7 @@
 import React, { Fragment, useEffect } from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import { getStudentProfileById } from "../../actions/profile";
+import { getStudentProfileById } from "../../../actions/profile";
 import { Container, Row, Col, Spinner } from "react-bootstrap";
 
 const Profile = ({
