+++ client/src/components/views/Unused/Photo.js
